update check style
