--- src/main/java/hu/progkorny/MetalCD/MetalCdApplication.java
@@ -1,13 +1,0 @@
-package hu.progkorny.MetalCD;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MetalCdApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MetalCdApplication.class, args);
-	}
-
-}
