--- src/main/java/hu/progkorny/MetalCD/model/Type.java
@@ -1,8 +1,0 @@
-package hu.progkorny.MetalCD.model;
-
-public enum Type {
-
-  MAXI,
-  MINI,
-  SINGLE
-}
