+++ src/main/java/hu/progkorny/metalcd/MetalCdApplication.java
@@ -1,0 +1,22 @@
+package hu.progkorny.metalcd;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+/**
+ * THIS IS MY METAL CD APPLICATIONS.
+ *
+ * @AUTHOR KULCSAR LASZLO ATTILA
+ * @VERSION 0.0.3
+ */
+@SpringBootApplication
+public class MetalCdApplication {
+
+    /**
+     * THIS IS MAIN METHOD.
+     */
+    public static void main(String[] args) {
+        SpringApplication.run(MetalCdApplication.class, args);
+    }
+
+}
