+++ src/main/java/hu/progkorny/metalcd/config/WebConfig.java
@@ -1,10 +1,13 @@
-package hu.progkorny.MetalCD.config;
+package hu.progkorny.metalcd.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+/**
+ * SPRING JAVA CONFIGURATION CLASS FOR WEB APPLICATION.
+ */
 @Configuration
 @EnableWebMvc
 public class WebConfig implements WebMvcConfigurer {
