+++ src/main/java/hu/progkorny/metalcd/controller/MetalCDController.java
@@ -1,9 +1,10 @@
-package hu.progkorny.MetalCD.controller;
+package hu.progkorny.metalcd.controller;
 
-import hu.progkorny.MetalCD.model.MetalCD;
-import hu.progkorny.MetalCD.model.exception.NotFoundException;
-import hu.progkorny.MetalCD.service.MetalCDService;
 import java.util.List;
+
+import hu.progkorny.metalcd.model.MetalCD;
+import hu.progkorny.metalcd.model.exception.NotFoundException;
+import hu.progkorny.metalcd.service.MetalCDService;
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -13,6 +14,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+/**
+ * CONTROLLER CLASS.
+ */
 @Controller
 @RequestMapping("/metalcd")
 public class MetalCDController {
@@ -27,6 +31,9 @@ public class MetalCDController {
     this.metalCDService = metalCDService;
   }
 
+  /**
+   * GET ALL METAL CDS.
+   */
   @GetMapping
   public String getAllMetalCD(final Model model) {
     final List<MetalCD> metalCDS = metalCDService.getAllMetalCDs();
@@ -34,6 +41,9 @@ public class MetalCDController {
     return METALCD_LIST_TEMPLATE_NAME;
   }
 
+  /**
+   * GET METAL CD ID.
+   */
   @GetMapping("/{id}")
   public String getMetalCD(final Model model, final @PathVariable Long id) {
     final MetalCD metalCD = metalCDService.getMetalCD(id);
@@ -42,6 +52,9 @@ public class MetalCDController {
   }
 
 
+  /**
+   * HOST UPDATE METAL CD.
+   */
   @PostMapping(value = "/update", consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
   public String updateMetalCD(final Model model,
                                final @RequestParam(value = "id", required = false) Long id,
@@ -51,11 +64,17 @@ public class MetalCDController {
     return METALCD_EDIT_TEMPLATE_NAME;
   }
 
+  /**
+   * GET CREATE METAL CD.
+   */
   @GetMapping("/create")
   public String createMetalCDForm(final Model model) {
     return "MetalCD/create";
   }
 
+  /**
+   * POST CREATE METAL CD.
+   */
   @PostMapping("/create")
   public String createMetalCD(final Model model, final MetalCD metalCD) {
     final MetalCD savedMetalCD = metalCDService.createMetalCD(metalCD);
@@ -63,6 +82,9 @@ public class MetalCDController {
     return METALCD_EDIT_TEMPLATE_NAME;
   }
 
+  /**
+   * DELETE METAL CD.
+   */
   @GetMapping("/{id}/delete")
   public String deleteMetalCD(final Model model, final @PathVariable("id") Long id) {
     try {
