+++ src/main/java/hu/progkorny/metalcd/model/Type.java
@@ -1,0 +1,11 @@
+package hu.progkorny.metalcd.model;
+
+/**
+ * TYPE ENUM.
+ */
+public enum Type {
+
+  MAXI,
+  MINI,
+  SINGLE
+}
