+++ src/main/resources/templates/metalcd/list.html
@@ -33,7 +33,7 @@
             </tr>
         </thread>
         <tbody>
-        <tr th:each="metalCD: ${metalCDS}">
+        <tr th:each="metalCD: ${metalcds}">
             <th scope="row"
                 th:text="${metalCD.id}"></th>
             <td th:text="${metalCD.name}"></td>
